refactor: modularize python.rs into organized module structure
Break down the monolithic python.rs (1,468 lines) into focused modules:

- src/python/mod.rs: Main module with re-exports and pymodule definition
- src/python/types.rs: All Python wrapper types and conversions
- src/python/analysis.rs: Core CSV/JSON analysis functions
- src/python/batch.rs: Batch processing and PyBatchAnalyzer
- src/python/ml.rs: ML readiness analysis and PyMlAnalyzer
- src/python/dataframe.rs: Pandas DataFrame integration
- src/python/logging.rs: Python logging system integration
- src/python/processor.rs: PyCsvProcessor context manager

This refactoring improves code maintainability while preserving all
existing Python API functionality. Part 2/3 of issue #52.
